// src/app/admin/enums/tabLabels.ts
export enum TabLabels {
    ADMIN = 'Admin profiles',
    UNIVERSITY = 'University profiles',
    STUDENT = 'Student profiles'
}

export enum ColumnHeadings {
    STUDENT = 'Student Name',
    ADMIN = 'Admin Name',
    EMAIL = 'Email',
    CELL = 'Cellphone Number',
    RANK = 'Rank',
    ROLE = 'Role'
}