// src/app/admin/enums/AdminProfileDialogButtons.ts
export enum AdminProfileDialogButtons {
    SEND_EMAIL = "Send registration email",
    ADD_DEPARTMENT = "Add department",
    UPDATE_DEPARTMENT = "Update department",
    REMOVE_DEPARTMENT = "Remove department",
    ADD_UNIVERSITY = "Add university",
    UPDATE_UNIVERSITY = "Update university details",
    REMOVE_UNIVERSITY = "Remove university",
    UPDATE_ADMIN = "Update admin details",
    UPDATE_STUDENT= "Update student details",
   
}